EduGuide/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
python-dotenv
PyMuPDF
requests
